# docs/swagger.yaml
definitions:
  errcode.Error:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      err:
        type: error
      msg:
        type: string
    type: object
...
  service.UserCreateRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  service.UserLoginRequest:
    properties:
      password:
        type: string
      rememberMe:
        type: boolean
      username:
        type: string
    required:
    - password
    - username
    type: object
  service.UserResponse:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      fansNums:
        type: integer
      gender:
        type: integer
      id:
        type: integer
      introduce:
        type: string
      root:
        type: boolean
      username:
        type: string
      words:
        type: integer
      workDay:
        type: integer
      worksNums:
        type: integer
    type: object
  service.UserSettingRequest:
    properties:
      commentEmail:
        type: boolean
      disableSearch:
        type: boolean
      hiddenGrade:
        type: boolean
      hiddenTag:
        type: boolean
      showAdult:
        type: boolean
      showEmail:
        type: boolean
      subscriptionEmail:
        type: boolean
      systemEmail:
        type: boolean
      topicEmail:
        type: boolean
    required:
    - commentEmail
    - disableSearch
    - hiddenGrade
    - hiddenTag
    - showAdult
    - showEmail
    - subscriptionEmail
    - systemEmail
    - topicEmail
    type: object
  service.UserSettingResponse:
    properties:
      commentEmail:
        type: boolean
      disableSearch:
        type: boolean
      hiddenGrade:
        type: boolean
      hiddenTag:
        type: boolean
      showAdult:
        type: boolean
      showEmail:
        type: boolean
      subscriptionEmail:
        type: boolean
      systemEmail:
        type: boolean
      topicEmail:
        type: boolean
    required:
    - commentEmail
    - disableSearch
    - hiddenGrade
    - hiddenTag
    - showAdult
    - showEmail
    - subscriptionEmail
    - systemEmail
    - topicEmail
    type: object
  service.UserUpdateEmailRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  service.UserUpdatePassRequest:
    properties:
      oldPassword:
        type: string
      password:
        type: string
    required:
    - oldPassword
    - password
    type: object
  service.UserUpdateRequest:
    properties:
      avatar:
        type: string
      gender:
        type: integer
      introduce:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: 属于国内的同人圈
  license: {}
  termsOfService: https://github.com/shyptr/archiveofourown
  title: 同人圈
  version: "1.0"
paths:
...
  /api/v1/currentUser:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.UserResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 当前用户信息
      tags:
      - 用户
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 修改用户信息
      tags:
      - 用户
  /api/v1/currentUser/email:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.UserUpdateEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 修改用户邮箱
      tags:
      - 用户
  /api/v1/currentUser/password:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.UserUpdatePassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 修改用户密码
      tags:
      - 用户
  /api/v1/currentUser/setting:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.UserSettingResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 当前用户偏好设置
      tags:
      - 用户
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.UserSettingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 修改用户偏好设置
      tags:
      - 用户
  /api/v1/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 登录
      tags:
      - 用户
  /api/v1/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 退出登录
      tags:
      - 用户
...
  /api/v1/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册邮箱
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 注册
      tags:
      - 用户
  /api/v1/register/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 创建用户
      tags:
      - 用户
  /api/v1/register/identify:
    get:
      consumes:
      - application/json
      parameters:
      - description: 注册路径
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 邮箱
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 验证注册邮箱
      tags:
      - 用户
...
  /api/v1/user/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.UserResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 用户信息
      tags:
      - 用户
swagger: "2.0"